move isfinite function to its op
