+++ paddle/fluid/framework/tensor_util.h
@@ -60,16 +60,6 @@ class PrintOptions {
   PrintOptions() {}
 };
 
-// store the result bool in gpu tensor, async operation. Faster than above ones.
-void TensorContainsNAN(const framework::Tensor& tensor, framework::Tensor* out);
-void TensorContainsInf(const framework::Tensor& tensor, framework::Tensor* out);
-void TensorIsfinite(const framework::Tensor& tensor, framework::Tensor* out);
-
-// copy the result bool to cpu
-bool TensorContainsNAN(const framework::Tensor& tensor);
-bool TensorContainsInf(const framework::Tensor& tensor);
-bool TensorIsfinite(const framework::Tensor& tensor);
-
 void TensorToStream(std::ostream& os,
                     const Tensor& tensor,
                     const platform::DeviceContext& dev_ctx);
